third_party/WebKit/LayoutTests/inspector-protocol/layers/paint-profiler.html

<html>
<head>
<style type="text/css">
.composited {
   position: absolute;
   top: 25px;
   left: 25px;
   width: 50px;
   height: 50px;
   background-color: blue;
   transform: translateZ(10px);
}
</style>
<script type="application/x-javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script type="application/x-javascript" src="../resources/layer-protocol-test.js"></script>
<script type="application/x-javascript">

function test()
{
    var documentNode;
    var initialLayers;
    var modifiedLayers;

    InspectorTest.enableLayerTreeAgent(gotLayerTree);

    function gotLayerTree(layers)
    {
        var matchingLayers = layers.filter(function(layer) { return !!(layer.backendNodeId && layer.transform); });
        InspectorTest.log("matchingLayers.length: " + matchingLayers.length);
        var layerId = matchingLayers[0].layerId;

        InspectorTest.sendCommand("LayerTree.makeSnapshot", {"layerId": layerId}, InspectorTest.wrapCallback(gotSnapshot));
    }
    var snapshotId;
    function gotSnapshot(result)
    {
        snapshotId = result.snapshotId;
        InspectorTest.sendCommand("LayerTree.profileSnapshot", {"snapshotId": snapshotId, "minRepeatCount":4, "minDuration": 0}, InspectorTest.wrapCallback(gotProfile));
    }
    function gotProfile(result)
    {
        var timings = result.timings;
        InspectorTest.log("Profile array length: " + result.timings.length);
        for (var i = 0; i < result.timings.length; ++i) {
            InspectorTest.log("Profile subarray " + i + " length: " + result.timings[i].length);
            for (var j = 0; j < result.timings[i].length; ++j)
                InspectorTest.log("Profile timing [" + i + "][" + j + "] is a number: " + (result.timings[i][j] >= 0));
        }
        InspectorTest.sendCommand("LayerTree.replaySnapshot", {"snapshotId": snapshotId, "fromStep": 2, "toStep": result.timings[0].length - 2}, InspectorTest.wrapCallback(replayedSnapshot));
    }
    function replayedSnapshot(result)
    {
        InspectorTest.log("LayerTree.replaySnapshot returned valid image: " + /^data:image\/png;base64,/.test(result.dataURL));
        InspectorTest.log("DONE!");
        InspectorTest.completeTest();
    }
}

</script>
<body onload="runTest()">
    <div class="composited">
    Sanity test for DevTools Paint Profiler.
    </div>
</body>
</html>
